Renamed vectordb as perth snakecase convention and modified relevant imports.
