+++ embedchain/config/__init__.py
@@ -8,8 +8,8 @@ from .embedder.base import BaseEmbedderConfig
 from .embedder.base import BaseEmbedderConfig as EmbedderConfig
 from .embedder.ollama import OllamaEmbedderConfig
 from .llm.base import BaseLlmConfig
-from .vectordb.chroma import ChromaDbConfig
-from .vectordb.elasticsearch import ElasticsearchDBConfig
-from .vectordb.opensearch import OpenSearchDBConfig
-from .vectordb.zilliz import ZillizDBConfig
+from .vector_db.chroma import ChromaDbConfig
+from .vector_db.elasticsearch import ElasticsearchDBConfig
+from .vector_db.opensearch import OpenSearchDBConfig
+from .vector_db.zilliz import ZillizDBConfig
 from .mem0_config import Mem0Config
