+++ embedchain/config/vector_db/lancedb.py
@@ -1,6 +1,6 @@
 from typing import Optional
 
-from embedchain.config.vectordb.base import BaseVectorDbConfig
+from embedchain.config.vector_db.base import BaseVectorDbConfig
 from embedchain.helpers.json_serializable import register_deserializable
