+++ tests/vectordb/test_lancedb.py
@@ -5,7 +5,7 @@ import pytest
 
 from embedchain import App
 from embedchain.config import AppConfig
-from embedchain.config.vectordb.lancedb import LanceDBConfig
+from embedchain.config.vector_db.lancedb import LanceDBConfig
 from embedchain.vectordb.lancedb import LanceDB
 
 os.environ["OPENAI_API_KEY"] = "test-api-key"
