+++ tests/vectordb/test_pinecone.py
@@ -1,6 +1,6 @@
 import pytest
 
-from embedchain.config.vectordb.pinecone import PineconeDBConfig
+from embedchain.config.vector_db.pinecone import PineconeDBConfig
 from embedchain.vectordb.pinecone import PineconeDB
