+++ tests/vectordb/test_qdrant.py
@@ -7,7 +7,7 @@ from qdrant_client.http.models import Batch
 
 from embedchain import App
 from embedchain.config import AppConfig
-from embedchain.config.vectordb.pinecone import PineconeDBConfig
+from embedchain.config.vector_db.pinecone import PineconeDBConfig
 from embedchain.embedder.base import BaseEmbedder
 from embedchain.vectordb.qdrant import QdrantDB
